+++ Connectors/CSi/Speckle.Connectors.CSiShared/HostApp/CSiSendCollectionManager.cs
@@ -4,6 +4,13 @@ using Speckle.Sdk.Models.Collections;
 
 namespace Speckle.Connectors.CSiShared.HostApp;
 
+/// <summary>
+/// We can use the CSiWrappers to create our collection structure.
+/// </summary>
+/// <remarks>
+/// This class manages the collections. If the key (from the path) already exists, this collection is returned.
+/// If it doesn't exist, a new collection is created and added to the rootObject.
+/// </remarks>
 public class CSiSendCollectionManager
 {
   private readonly IConverterSettingsStore<CSiConversionSettings> _converterSettings;
@@ -14,7 +21,7 @@ public class CSiSendCollectionManager
     _converterSettings = converterSettings;
   }
 
-  // TODO: Frames could be further classified under Columns, Braces and Beams. Same for shells: walls, floors
+  // TODO: Frames could be further classified under Columns, Braces and Beams. Same for Shells which could be classified into walls, floors
   public Collection GetAndCreateObjectHostCollection(ICSiWrapper csiObject, Collection rootObject)
   {
     var path = csiObject.GetType().Name.Replace("Wrapper", ""); // CSiJointWrapper → CSiJoint, CSiFrameWrapper → CSiFrame etc.
