+++ Connectors/CSi/Speckle.Connectors.CSiShared/Speckle.Connectors.CSiShared.projitems
@@ -15,14 +15,14 @@
     <Compile Include="$(MSBuildThisFileDirectory)Filters\CSiSharedSelectionFilter.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)HostApp\CSiSendCollectionManager.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Operations\Send\CSiRootObjectBuilder.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Plugin\CSiSharedPluginBase.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Plugin\CSiPluginBase.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Plugin\SpeckleFormBase.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="$(MSBuildThisFileDirectory)GlobalUsing.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)HostApp\CSiSharedApplicationService.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)HostApp\CSiSharedDocumentModelStore.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)HostApp\CSiSharedIdleManager.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)HostApp\CSiApplicationService.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)HostApp\CSiDocumentModelStore.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)HostApp\CSiIdleManager.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ServiceRegistration.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Utils\ObjectIdentifiers.cs" />
   </ItemGroup>
