+++ Converters/CSi/Speckle.Converters.CSiShared/CSiWrappers.cs
@@ -6,6 +6,14 @@ public interface ICSiWrapper
   int ObjectType { get; }
 }
 
+/// <summary>
+/// Based on GetSelected() returns of objectType and objectName, we need to create a CSiWrapper object.
+/// </summary>
+/// <remarks>
+/// Creating a class that can be used to pass a type to the converter.
+/// Since the API only provides a framework for us to query the model, we don't get instances.
+/// The types are the same for both SAP 2000 and ETABS.
+/// </remarks>
 public abstract class CSiWrapperBase : ICSiWrapper
 {
   public required string Name { get; set; }
@@ -47,6 +55,13 @@ public class CSiLinkWrapper : CSiWrapperBase
   public override int ObjectType => 7;
 }
 
+/// <summary>
+/// ObjectType specific wrappers created during bindings.
+/// </summary>
+/// <remarks>
+/// Switch statements based off of the objectType int return.
+/// Used in the connectors and allows converters to be resolved effectively.
+/// </remarks>
 public static class CSiWrapperFactory
 {
   public static ICSiWrapper Create(int objectType, string name) =>
